pyproject.toml
[tool.poetry]
name = "kasa-monitor"
version = "0.1.0"
description = ""
authors = ["c8763yee <rrrrr4413@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "kasa_monitor" }]

[tool.poetry.dependencies]
python = "^3.10"
python-kasa = { extras = ["shell", "speedups"], version = "*" }
paho-mqtt = "^2.1.0"
alembic = "1.13.1"
greenlet = "3.0.3"
mako = "1.3.2"
markupsafe = "2.1.5"
mysqlclient = "2.2.4"
python-dotenv = "1.0.1"
sqlalchemy = "2.0.28"
typing-extensions = "4.10.0"
pydantic = "^2.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
